requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=8.0.0
wandb>=0.15.0
tqdm>=4.65.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
albumentations>=1.3.0
insightface>=0.7.0  # For ArcFace embeddings 